ПР 13.1 fix 1

+++ controllers/card.js
@@ -1,3 +1,4 @@
+const { isValidObjectId } = require("mongoose");
 const Card = require("../models/card");
 
 module.exports.getCards = (req, res) => {
@@ -19,28 +20,36 @@ module.exports.getCards = (req, res) => {
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
   const owner = req.user._id;
-  if (!name || !link || !owner)
-    return res.status(400).send({
-      message: "Переданы некорректные данные для создания карточки",
-    });
+
   Card.create({ name, link, owner })
     .then(({ _id, name, link, owner, likes }) =>
       res.send({ _id, name, link, owner, likes })
     )
-    .catch((err) => res.status(500).send({ message: err.message }));
+    .catch((err) => {
+      if (err.name === "ValidationError") {
+        res.status(400).send({ message: "Переданы некорректные данные" });
+      } else {
+        res.status(500).send({ message: err.message });
+      }
+    });
 };
 
 module.exports.removeCard = (req, res) => {
   const { cardId } = req.params;
 
+  if (!isValidObjectId(cardId))
+    return res.status(400).send({
+      message: "Переданы некорректные данные для удаления карточки",
+    });
+
   Card.deleteOne({ _id: cardId })
-    .then(() => res.status(200).send())
+    .then(({ deletedCount }) => {
+      if (!deletedCount)
+        return res.status(404).send({ message: "Карточка не найдена" });
+      res.status(200).send({ message: "Карточка удалена" });
+    })
     .catch((err) => {
-      if (err.name === "CastError") {
-        res.status(404).send({ message: "Карточка не найдена" });
-      } else {
-        res.status(500).send({ message: err.message });
-      }
+      res.status(500).send({ message: err.message });
     });
 };
 
@@ -48,7 +57,7 @@ module.exports.addCardLike = (req, res) => {
   const { cardId } = req.params;
   const userId = req.user._id;
 
-  if (!userId)
+  if (!isValidObjectId(userId) || !isValidObjectId(cardId))
     return res.status(400).send({
       message: "Переданы некорректные данные для постановки лайка",
     });
@@ -58,13 +67,13 @@ module.exports.addCardLike = (req, res) => {
     { $addToSet: { likes: userId } },
     { new: true }
   )
-    .then((card) => res.send(card.likes))
+    .then((card) => {
+      if (!card)
+        return res.status(404).send({ message: "Карточка не найдена" });
+      res.send(card.likes);
+    })
     .catch((err) => {
-      if (err.name === "CastError") {
-        res.status(404).send({ message: "Карточка не найдена" });
-      } else {
-        res.status(500).send({ message: err.message });
-      }
+      res.status(500).send({ message: err.message });
     });
 };
 
@@ -72,18 +81,18 @@ module.exports.removeCardLike = (req, res) => {
   const { cardId } = req.params;
   const userId = req.user._id;
 
-  if (!userId)
+  if (!isValidObjectId(userId) || !isValidObjectId(cardId))
     return res.status(400).send({
       message: "Переданы некорректные данные для снятия лайка",
     });
 
   Card.findByIdAndUpdate(cardId, { $pull: { likes: userId } }, { new: true })
-    .then((card) => res.send(card.likes))
+    .then((card) => {
+      if (!card)
+        return res.status(404).send({ message: "Карточка не найдена" });
+      res.send(card.likes);
+    })
     .catch((err) => {
-      if (err.name === "CastError") {
-        res.status(404).send({ message: "Карточка не найдена" });
-      } else {
-        res.status(500).send({ message: err.message });
-      }
+      res.status(500).send({ message: err.message });
     });
 };
